update: add types
